feat: add client-side routing for SEO
- Add react-router-dom for client-side navigation
- Create separate pages: LandingPage (/), AppPage (/app), AboutPage (/about)
- Landing page shows entry screen with Begin button
- App page auto-initializes audio experience
- About page explains the concept with og-image
