+++ package.json
@@ -15,6 +15,7 @@
     "@vercel/speed-insights": "^1.3.1",
     "react": "^19.2.1",
     "react-dom": "^19.2.1",
+    "react-router-dom": "^7.10.1",
     "tone": "^15.1.22"
   },
   "devDependencies": {
